requirements_old.txt
asgiref==3.7.2
Django==4.2.5
django-channels==0.7.0
django-imagekit==5.0.0
sqlparse==0.4.4
tzdata==2023.3

none~=0.1.1
null~=0.6.1
channels~=4.0.0
asgiref~=3.7.2

